feat: Enhance admin functionality with category and product management routes and controllers

+++ src/util/cloudinary.js
@@ -1,5 +1,7 @@
 import { v2 as cloudinary } from "cloudinary";
+import dotenv from "dotenv"
 import fs from "fs";
+dotenv.config();
 cloudinary.config({
   cloud_name: process.env.CLOUDINARY_NAME,
   api_key: process.env.CLOUDINARY_API_KEY,
@@ -14,9 +16,14 @@ const uploadCloudinary = async (localPath) => {
       resource_type: "auto",
     });
     console.log("file is uploaded in cloudinary", response.url);
+     if (fs.existsSync(localPath)) {
+       fs.unlinkSync(localPath);
+     }
     return response;
   } catch (err) {
-    fs.unlinkSync(localPath);
+     if (fs.existsSync(localPath)) {
+       fs.unlinkSync(localPath);
+     }
     return null;
   }
 };
